Fix version number from being too demanding
1.25.0 is too demanding
While I was testing with older versions of go While
importing auth, the moment I stepped below 1.25.0
I got an error from the compiler mentioning that
auth needs 1.25.0

For comparision our packages (the ones we imported)
the most demanding one is pgx.
i.e 1.24.0

Note that Go lang doesn't allow older versions to
go on with compilations. If the go.mod defines
1.25.0 then the version of your compiler would have
to be >= 1.25.0, not less than that.

So we bring it down to pgx equivalant that is
1.24.0 while maintaining a good version with others too.

Signed-off-by: K Jayatheerth <jayatheerthkulkarni2005@gmail.com>